requirements-dev.txt
# Requirements de développement pour ICT Group Website
# Inclut requirements.txt plus les outils de développement

-r requirements.txt

# Outils de développement
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0
ipython>=8.10.0
ipdb>=0.13.13

# Tests
pytest>=7.2.0
pytest-django>=4.5.2
pytest-cov>=4.0.0
factory-boy>=3.2.1

# Linting et formatage
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pre-commit>=3.0.0

# Documentation
Sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Monitoring local
django-silk>=5.0.0  # Profiling des requêtes
memory-profiler>=0.60.0

# Outils de base de données
django-migration-testcase>=1.0.0
